Iniital import of SyncPrims
This commit contains the initial version of the Synchronization Primitives (SyncPrims) microbenchmarks.  They are based on work by Stuart and Owens, and have been modified according the our IISWC '17 paper.  A subsequent commit will add the relaxed atomics microbenchmarks from the paper.

There are two versions of the microbenchmarks in this commit:

1.  The denovo version (syncPrims/cuda/denovo/), which is the version we ran for the IISWC '17 paper with only minor changes to comment out the DeNovo specific code.  This version was tested with CUDA 3.2, because this was the last version of CUDA fully supported by GPGPU-Sim (at the time).

2.  A CUDA UVM version.  This was tested on a Pascal P100 GPU with CUDA 6.0 installed.  The main difference from the DeNovo version is that this version runs on real GPUs, not just the DeNovo simulator.