fix golden code for local Semaphores in HeteroSync
Like the golden code for the global semaphores in HeteroSync, the
golden code for the local semaphore needs to be separate from the
local mutex calculations because the writer WG per CU accesses more
and different data than the mutexes do. However, unlike the global
semaphore we can model this behavior by iterating over all of the
data that will be accessed across all CUs, since it has the same
effect as long as the number of WGs/CU are the same.

Change-Id: I06b67a4ca3ec313d6f9568949a6cc76457910a97