hip: replace backoff loops with s_sleep
GCN3 has an instruction for sleeping threads/WGs (s_sleep), which
is more efficient than the dummy loops we were using because it
doesn't require any resources (i.e., it doesn't busy-wait).  Using
this instead of the busy-wait loops improves performance for the
benchmarks that use waiting.  Moreover, I used a block of if's
that decrease sleep amounts by powers of 2 because this was roughly
5X faster than a loop that did s_sleep(1) repeatedly.  I believe
this is because the longer sleeps amortize the overhead of
sleeping (and this outweights the branch divergence my version
causes).