cuda: add nanosleep functionality
For all microbenchmarks that use backoff, add option to use
CUDA's new (as of Volta-class GPUs) nanosleep built-in intrinsic.
This does backoff using CUDA's nanosleep instead of strawman
NOP loops, which were required for backoff in previous CUDA
versions.

To ensure backwards compatibility and avoid code bloat, to enable
nanosleep, this patch adds a flag in the Makefile.  This flag
should be set by the user (either on the command line or in the
Makefile) when using a Volta-class or newer NVIDIA GPU and they
want to use nanosleep instead of NOP loops for backoff.  In the
code, alongside this new flag (HAS_NANOSLEEP) we check if the
version of CUDA (11+) provides nanosleep support.

Using nanosleep instead of NOP loops provides a small (< 5%)
performance benefit, although further work could be done to try
and find the optimal backoff level for nanosleep.