uts: CUDA UVM version of UTS.
Takes the DeNovo version of UTS and converts to use CUDA UVM
through converting cudaMalloc's to cudaMallocManaged's, adding
threadfence's at the appropriate places, adding a
cudaDeviceSynchronize after the kernel is called, and updating
the Makefile appropriately to run for newer GPUs.  This has been
tested with CUDA 11.1 on a GTX1080 for the size in the run file.
Additional testing needed for newer and larger GPU, and updating
the Makefile appropriately to run for newer GPUs.  This has been
tested with CUDA 11.1 on a GTX1080 for the size in the run file.
Additional testing needed for newer and larger GPUs.